Uploaded car.hs
I demonstrate object oriented programming in Haskell using my car blueprint as in some of my other projects. Haskell values are immutable, and so clones must be formed. The functions used are essentially constructors that take an object and replicate it, though slightly modified. This is as close as it gets to OOP in Haskell.